+++ main-app/pages/_app.tsx
@@ -3,12 +3,12 @@ import { SessionContextProvider, Session } from "@supabase/auth-helpers-react";
 import type { AppProps } from "next/app";
 import { ReactElement, ReactNode, useState } from "react";
 import { NextPage } from "next";
-import { Layout } from "../components/Layout";
+import { Layout } from "components/Layout";
 import { NextUIProvider } from "@nextui-org/react";
 import { SSRProvider } from "@react-aria/ssr";
 import Head from "next/head";
-import ConfirmContextProvider from "../providers/ConfirmContextProvider";
-import { ConfirmModal } from "../components/modals/ConfirmModal";
+import ConfirmContextProvider from "providers/ConfirmContextProvider";
+import { ConfirmModal } from "components/modals/ConfirmModal";
 import { appWithTranslation } from "next-i18next";
 
 export type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
