+++ main-app/pages/api/translate.ts
@@ -1,26 +1,26 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { translateWord } from "../../util/azure/translation/translation";
+import { translateWord } from "utils/azure/translation/translation";
 
 type Data = {
-	translatedWord: string;
+  translatedWord: string;
 };
 
 export default async function handler(
-	req: NextApiRequest,
-	res: NextApiResponse<Data>
+  req: NextApiRequest,
+  res: NextApiResponse<Data>
 ) {
-	const { method } = req;
-	switch (method) {
-		case "POST":
-			const { word, targetLanguage, sourceLanguage } = req.body;
-			if (!word || !targetLanguage || !sourceLanguage) {
-				throw new Error("invalid props");
-			}
-			const translation = await translateWord(
-				word,
-				sourceLanguage,
-				targetLanguage
-			);
-			res.status(200).send({ translatedWord: translation });
-	}
+  const { method } = req;
+  switch (method) {
+    case "POST":
+      const { word, targetLanguage, sourceLanguage } = req.body;
+      if (!word || !targetLanguage || !sourceLanguage) {
+        throw new Error("invalid props");
+      }
+      const translation = await translateWord(
+        word,
+        sourceLanguage,
+        targetLanguage
+      );
+      res.status(200).send({ translatedWord: translation });
+  }
 }
