+++ main-app/pages/app/texts/index.tsx
@@ -1,15 +1,15 @@
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { Container, red, Row, Table, Tooltip } from "@nextui-org/react";
-import { TextRow } from "../../../types/Texts";
-import { getRouteForSingleText } from "../../../util/routing/texts";
+import { TextRow } from "types/Texts";
+import { getRouteForSingleText } from "utils/routing/texts";
 import { GetServerSidePropsContext } from "next";
 import { createServerSupabaseClient } from "@supabase/auth-helpers-nextjs";
-import { Database } from "../../../types/supabase";
-import { getLocale } from "../../../util/translation/getLocale";
+import { Database } from "types/supabase";
+import { getLocale } from "utils/translation/getLocale";
 import { useTranslation } from "next-i18next";
-import { FullTable } from "../../../components/table/FullTable";
-import { useConfirm } from "../../../hooks/useConfirm";
-import { deleteText, getTexts } from "../../../util/supabase/texts";
+import { FullTable } from "components/table/FullTable";
+import { useConfirm } from "hooks/useConfirm";
+import { deleteText, getTexts } from "utils/supabase/texts";
 import { useSupabaseClient } from "@supabase/auth-helpers-react";
 import { useState } from "react";
 import { useRouter } from "next/router";
