+++ main-app/pages/login/index.tsx
@@ -1,6 +1,6 @@
 import { useUser } from "@supabase/auth-helpers-react";
 import React from "react";
-import { LoginForm } from "../../components/authentication/LoginForm";
+import { LoginForm } from "components/authentication/LoginForm";
 
 export default function Login() {
   const user = useUser();
