+++ main-app/utils/mapping/flashcards.ts
