+++ main-app/utils/translation/getLocale.ts
